Upgrade to Kibana 4 and kibana_lwrp
- Move from Berksfile entry to kibana_lwrp on Supermarket. You should remove references to the kibana cookbook in all Berksfile files in any cookbooks that wrap elkstack, fixes #71
- Using the `file` installation method of kibana, which now includes a JVM as well, as Kibana 4 is a daemonized process instead of simply Javascript.
- Adjusted tests to reflect Kibana 4's successful output on a page load

Some additional tuning may be needed to configure kibana beyond the changes here, so that another JVM will co-exist with the Logstash and Elasticsearch JVMs (this is getting out of hand) but this is a first pass. We may want to consider running kibana on a separate server as a supported configuration for elkstack in the future.